+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { Router } from "wouter";
 import { useHashLocation } from "wouter/use-hash-location";
-import App from "./components/app";
+import App from "./components/app/app";
 import { Provider } from "./components/ui/provider";
 import "./index.css";